requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.12.0
pycparser==2.21
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
